% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{breathe}
\alias{breathe}
\title{breathe}
\usage{
breathe(color, from_color = current_color(selector)[[1]], period = 10,
  cycles = 2, persist = FALSE, peak = 0.5, selector = "all",
  power_on = TRUE)
}
\arguments{
\item{color}{a string describing the desired color; see examples.}

\item{from_color}{Same syntax as color, defaults to current color (of first bulb in selection)}

\item{period}{time in seconds for the cycle to take place}

\item{cycles}{number of cycles to perform}

\item{persist}{should the color persist after the effect? default is FALSE (returns to original color)}

\item{peak}{when in the cycle should the color be at it's maximum intensity?}

\item{selector}{a string in format '[type]:[value]', where type can be
'all', 'id', 'label', 'group', 'group_id', 'location', 'location_id',
'scene_id', and value is what you want to target. The default is 'all',
which needs no value argument.}

\item{power_on}{should the light be powered on if it is off? (FALSE just leaves light off)}
}
\description{
Display a breathe effect
}
\examples{
\dontrun{
breathe("purple", "blue")
}
}

