% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{color}
\alias{color}
\title{color}
\usage{
color(color, selector = "all", duration = 1, power_on = TRUE)
}
\arguments{
\item{color}{a string describing the desired color; see examples.}

\item{selector}{a string in format '[type]:[value]', where type can be
'all', 'id', 'label', 'group', 'group_id', 'location', 'location_id',
'scene_id', and value is what you want to target. The default is 'all',
which needs no value argument.}

\item{duration}{the length of the effect}

\item{power_on}{should the light be powered on if it is off? (FALSE just leaves light off)}
}
\value{
httr response object
}
\description{
set lifx color
}
\details{
Note that the kelvin temperature ranges from 2700 to 8000. Hue in HSB list is
a number between 0 and 360, whereas saturation and brightness should be between 0 and 1.
}
\examples{
\dontrun{
 color("green", "label:desk")   # deep green, brightness untouched on lights labeled 'desk'
 color("blue brightness:100\%")  # deep blue, maximum brightness
 color("hsb:0,1,1")             # deep red, maximum brightness
 color("random")                # random hue, maximum saturation, brightness untouched
 color("kelvin:2700")           # warm white, brightness untouched
 color("saturation:100\%")       # set maximum saturation
}
}

