% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{current_color}
\alias{current_color}
\title{current color}
\usage{
current_color(selector = "all")
}
\arguments{
\item{selector}{a string in format '[type]:[value]', where type can be
'all', 'id', 'label', 'group', 'group_id', 'location', 'location_id',
'scene_id', and value is what you want to target. The default is 'all',
which needs no value argument.}
}
\value{
httr response object
}
\description{
list of the current colors, in a format valid for \code{\link{color}}
}

