% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{get_scenes}
\alias{get_scenes}
\title{Lists all the scenes available in the users account}
\usage{
get_scenes()
}
\value{
httr response object
}
\description{
Lists all the scenes available in the users account
}
