% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{label}
\alias{label}
\title{label}
\usage{
label(label, selector)
}
\arguments{
\item{label}{the label for the bulb}

\item{selector}{selector pattern for a single bubl, e.g. id:<idstring>}
}
\value{
httr response object
}
\description{
add a label to a bulb
}

