% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{parse_color}
\alias{parse_color}
\title{parse color}
\usage{
parse_color(string)
}
\arguments{
\item{string}{The color string to parse}
}
\value{
hsbk information for the string. NOTE: This API endpoing appears not to be working yet!
}
\description{
Parse a color string and return hue, saturation, brightness and kelvin values
}

