% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{pulse}
\alias{pulse}
\title{pulse}
\usage{
pulse(color, from_color = current_color(selector)[[1]], period = 5,
  cycles = 1, persist = FALSE, duty_cycle = 0.5, selector = "all",
  power_on = TRUE)
}
\arguments{
\item{color}{a string describing the desired color; see examples.}

\item{from_color}{Same syntax as color, defaults to current color (of first bulb in selection)}

\item{period}{time in seconds for the cycle to take place}

\item{cycles}{number of cycles to perform}

\item{persist}{should the color persist after the effect? default is FALSE (returns to original color)}

\item{duty_cycle}{Ratio of the period where color is active. Only used for pulse. Defaults to 0.5. Range: 0-1}

\item{selector}{a string in format '[type]:[value]', where type can be
'all', 'id', 'label', 'group', 'group_id', 'location', 'location_id',
'scene_id', and value is what you want to target. The default is 'all',
which needs no value argument.}

\item{power_on}{should the light be powered on if it is off? (FALSE just leaves light off)}
}
\value{
httr response object
}
\description{
pulse a color for a defined period
}

