% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{rate_limit_warner}
\alias{rate_limit_warner}
\title{Checks headers of incoming message for rate limiting
will complain if rate limit is less then 5. 
when it reaches 0 it will tell you when you can try again.
For internal use.}
\usage{
rate_limit_warner(results)
}
\arguments{
\item{results}{a httr response object}
}
\description{
Checks headers of incoming message for rate limiting
will complain if rate limit is less then 5. 
when it reaches 0 it will tell you when you can try again.
For internal use.
}
