% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{scene}
\alias{scene}
\title{scene}
\usage{
scene(state = c("on", "off"), scene_id, duration = 1)
}
\arguments{
\item{state}{on or off?}

\item{scene_id}{the id of the desired scene}

\item{duration}{the length of the effect}
}
\value{
httr response object
}
\description{
turn on a scene for a bulb
}

