% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{set_state}
\alias{set_state}
\title{Set the state, a general function}
\usage{
set_state(selector = "all", power = "on", color = NULL,
  brightness = NULL, saturation = NULL, duration = 1)
}
\arguments{
\item{selector}{Which lights do you want? defaults to "all"}

\item{power}{either on or off, defaults to on, makes not sense to shut the light off and change the color}

\item{color}{a name or combination of hue, brightness, saturation etc.}

\item{brightness}{between 0-1 how much light must the bulb emit}

\item{saturation}{between 0-1 how much color do you want to add to white}

\item{duration}{a fade duration in seconds, f.i. 1 second}
}
\description{
Set the state, a general function
}
