% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifx-cloud.R
\name{parse_color}
\alias{parse_color}
\title{parse color}
\usage{
parse_color(string)
}
\arguments{
\item{string}{The color string to parse}
}
\value{
hsbk information for the string.
}
\description{
Parse a color string and return hue, saturation, brightness and kelvin values
}
